refactor: rename playerDicisions => playerDecisions & delete unnecessary comment
